+++ CareMI/src/main/java/br/com/mapped/CareMI/dto/PacientePlanoSaudeDto/AtualizacaoPacientePlanoSaudeDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public record AtualizacaoPacientePlanoSaudeDto(
         @NotNull(message = "O número da carteira não pode ser nulo")
         @Positive(message = "O número da carteira deve ser um número positivo")
-        Integer carteira,
+        Long carteira,
 
         @NotNull(message = "A data de início não pode ser nula")
         LocalDate dataInicio,
