+++ CareMI/src/main/java/br/com/mapped/CareMI/dto/PacientePlanoSaudeDto/DetalhesPacientePlanoSaudeDto.java
@@ -3,7 +3,7 @@ import br.com.mapped.CareMI.model.PacientePlanoSaude;
 
 import java.time.LocalDate;
 
-public record DetalhesPacientePlanoSaudeDto(Long id, Integer carteira, LocalDate dataInicio, LocalDate dataFim) {
+public record DetalhesPacientePlanoSaudeDto(Long id, Long carteira, LocalDate dataInicio, LocalDate dataFim) {
 
     public DetalhesPacientePlanoSaudeDto(PacientePlanoSaude pacientePlanoSaude) {
         this(pacientePlanoSaude.getId(), pacientePlanoSaude.getCarteira(), pacientePlanoSaude.getDataInicio(), pacientePlanoSaude.getDataFim());
